requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio>=3.7.1,<4.0.0
faiss-cpu==1.7.4
asgiref==3.8.1
billiard==4.2.1
celery==5.3.6
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.2
cycler==0.12.1
Django==5.2
fastapi==0.104.1
starlette>=0.27.0,<0.28.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
image==1.5.33
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
kombu==5.5.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.3
packaging==25.0
pandas==2.2.3
pillow==10.0.0
prompt_toolkit==3.0.51
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.1
tenacity==8.2.3
threadpoolctl==3.6.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.24.0
vine==5.1.0
wcwidth==0.2.13
redis==5.2.1
aiohttp==3.9.5
requests==2.31.0
python-multipart==0.0.6
